+++ CHANGELOG.md
@@ -21,3 +21,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initiated `CHANGELOG.md`
+
+## [0.0.1.2] - 2024-05-24
+
+### Versions
+
+- CSC FastAPI Backend:  0.1.0.1
+- CSC React Frontend:   0.1.4.1
+- CSC Sheetscan Module: 0.0.1.10
+
+### Fixed
+
+#### CSC FastAPI Backend
+
+- Updated `ComponentModel` in `models.py`
+- Fixed response models in `routers.py`
+
+#### CSC React Frontend
+
+- Fixed timestamp printing in console
+
+### Added
+
+#### CSC React Frontend
+
+- Added `timestamp_string` function to `lib/utils.ts`